- handle close button - resolve bug on tng payment
