+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdk 34
 
     defaultConfig {
-        applicationId "com.molpay.molpayxdkproject"
+        applicationId "com.fiuu.xdkandroid"
         minSdkVersion 21
         targetSdkVersion 34
         versionCode 1
@@ -17,7 +17,7 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    namespace 'com.molpay.molpayxdkproject'
+    namespace 'com.fiuu.xdkandroid'
 
     viewBinding {
         enabled = true
@@ -26,9 +26,9 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.google.android.gms:play-services-wallet:19.3.0'
+    implementation 'com.google.android.gms:play-services-wallet:19.4.0'
     testImplementation 'junit:junit:4.13.2'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.12.0'
-    implementation project(':molpayxdk')
+    implementation project(':paymentxdk')
 }
