--- molpayxdk/build.gradle
@@ -1,55 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'maven-publish'
-}
-
-android {
-    compileSdk 34
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 34
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility(JavaVersion.VERSION_1_8)
-        targetCompatibility(JavaVersion.VERSION_1_8)
-    }
-
-    viewBinding {
-        enabled = true
-    }
-}
-
-dependencies {
-    testImplementation 'junit:junit:4.13.2'
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.12.0'
-
-    // Google Pay
-    implementation 'com.google.zxing:core:3.5.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation "androidx.lifecycle:lifecycle-viewmodel:2.7.0"
-    implementation "com.google.android.gms:play-services-pay:16.4.0"
-    implementation "com.google.android.gms:play-services-wallet:19.3.0"
-    implementation 'org.apache.commons:commons-lang3:3.4'
-}
-
-afterEvaluate {
-    publishing {
-        publications {
-            release(MavenPublication) {
-                from components.release
-                groupId ='com.github.FiuuPayment'
-                artifactId = 'Mobile-XDK-Fiuu_Android_Library'
-                version = '3.34.2'
-            }
-        }
-    }
-}
