--- molpayxdk/src/main/java/com/molpay/molpayxdk/googlepay/UtilGP.java
@@ -1,247 +1,0 @@
-package com.molpay.molpayxdk.googlepay;
-
-import android.content.Context;
-
-import com.google.android.gms.wallet.PaymentsClient;
-import com.google.android.gms.wallet.Wallet;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-
-/**
- * Contains helper static methods for dealing with the Payments API.
- *
- * <p>Many of the parameters used in the code are optional and are set here merely to call out their
- * existence. Please consult the documentation to learn more and feel free to remove ones not
- * relevant to your implementation.
- */
-public class UtilGP {
-
-    public static final BigDecimal CENTS_IN_A_UNIT = new BigDecimal(100);
-
-    /**
-     * Create a Google Pay API base request object with properties used in all requests.
-     *
-     * @return Google Pay API base request object.
-     * @throws JSONException if the object is malformed.
-     */
-    private static JSONObject getBaseRequest() throws JSONException {
-        return new JSONObject().put("apiVersion", 2).put("apiVersionMinor", 0);
-    }
-
-    /**
-     * Creates an instance of {@link PaymentsClient} for use in an {@link Context} using the
-     * environment and theme set in {@link ConstantsGP}.
-     *
-     * @param context is the caller's context.
-     */
-    public static PaymentsClient createPaymentsClient(Context context) {
-        Wallet.WalletOptions walletOptions =
-                new Wallet.WalletOptions.Builder().setEnvironment(ActivityGP.PAYMENTS_ENVIRONMENT).build();
-        return Wallet.getPaymentsClient(context, walletOptions);
-    }
-
-    /**
-     * Gateway Integration: Identify your gateway and your app's gateway merchant identifier.
-     *
-     * <p>The Google Pay API response will return an encrypted payment method capable of being charged
-     * by a supported gateway after payer authorization.
-     *
-     * <p>NOTE: molpay is RMS payment gateway for Google Pay
-     *
-     * @return Payment data tokenization for the CARD payment method.
-     * @throws JSONException if the object is malformed.
-     * @see <a href=
-     * "https://developers.google.com/pay/api/android/reference/object#PaymentMethodTokenizationSpecification">PaymentMethodTokenizationSpecification</a>
-     */
-    private static JSONObject getGatewayTokenizationSpecification() throws JSONException {
-        return new JSONObject() {{
-            put("type", "PAYMENT_GATEWAY");
-            put("parameters", new JSONObject() {{
-                put("gateway", "molpay");
-                put("gatewayMerchantId", "molpay");
-            }});
-        }};
-    }
-
-    /**
-     * Card networks supported by your app and your gateway.
-     *
-     * @return Allowed card networks
-     * @see <a
-     * href="https://developers.google.com/pay/api/android/reference/object#CardParameters">CardParameters</a>
-     */
-    private static JSONArray getAllowedCardNetworks() {
-        return new JSONArray(ConstantsGP.SUPPORTED_NETWORKS);
-    }
-
-    /**
-     * Card authentication methods supported by your app and your gateway.
-     *
-     * and make updates in Constants.java.
-     *
-     * @return Allowed card authentication methods.
-     * @see <a
-     * href="https://developers.google.com/pay/api/android/reference/object#CardParameters">CardParameters</a>
-     */
-    private static JSONArray getAllowedCardAuthMethods() {
-        return new JSONArray(ConstantsGP.SUPPORTED_METHODS);
-    }
-
-    /**
-     * Describe your app's support for the CARD payment method.
-     *
-     * <p>The provided properties are applicable to both an IsReadyToPayRequest and a
-     * PaymentDataRequest.
-     *
-     * @return A CARD PaymentMethod object describing accepted cards.
-     * @throws JSONException if the object is malformed.
-     * @see <a
-     * href="https://developers.google.com/pay/api/android/reference/object#PaymentMethod">PaymentMethod</a>
-     */
-    private static JSONObject getBaseCardPaymentMethod() throws JSONException {
-        JSONObject cardPaymentMethod = new JSONObject();
-        cardPaymentMethod.put("type", "CARD");
-
-        JSONObject parameters = new JSONObject();
-        parameters.put("allowedAuthMethods", getAllowedCardAuthMethods());
-        parameters.put("allowedCardNetworks", getAllowedCardNetworks());
-
-        // Optionally, you can add billing address/phone number associated with a CARD payment method.
-        parameters.put("billingAddressRequired", false);
-//    JSONObject billingAddressParameters = new JSONObject();
-//    billingAddressParameters.put("format", "FULL");
-//    parameters.put("billingAddressParameters", billingAddressParameters);
-
-        cardPaymentMethod.put("parameters", parameters);
-
-        return cardPaymentMethod;
-    }
-
-    /**
-     * Describe the expected returned payment data for the CARD payment method
-     *
-     * @return A CARD PaymentMethod describing accepted cards and optional fields.
-     * @throws JSONException if the object is malformed.
-     * @see <a
-     * href="https://developers.google.com/pay/api/android/reference/object#PaymentMethod">PaymentMethod</a>
-     */
-    private static JSONObject getCardPaymentMethod() throws JSONException {
-        JSONObject cardPaymentMethod = getBaseCardPaymentMethod();
-        cardPaymentMethod.put("tokenizationSpecification", getGatewayTokenizationSpecification());
-
-        return cardPaymentMethod;
-    }
-
-    /**
-     * Return a collection of payment methods allowed to complete the operation with Google Pay.
-     * @return A JSONArray object with the list of payment methods.
-     * @throws JSONException
-     */
-    public static JSONArray getAllowedPaymentMethods() throws JSONException {
-        return new JSONArray() {{
-            put(getCardPaymentMethod());
-        }};
-    }
-
-    /**
-     * An object describing accepted forms of payment by your app, used to determine a viewer's
-     * readiness to pay.
-     *
-     * @return API version and payment methods supported by the app.
-     * @see <a
-     * href="https://developers.google.com/pay/api/android/reference/object#IsReadyToPayRequest">IsReadyToPayRequest</a>
-     */
-    public static JSONObject getIsReadyToPayRequest() {
-        try {
-            JSONObject isReadyToPayRequest = getBaseRequest();
-            isReadyToPayRequest.put(
-                    "allowedPaymentMethods", new JSONArray().put(getBaseCardPaymentMethod()));
-
-            return isReadyToPayRequest;
-
-        } catch (JSONException e) {
-            return null;
-        }
-    }
-
-    /**
-     * Provide Google Pay API with a payment amount, currency, and amount status.
-     *
-     * @return information about the requested payment.
-     * @throws JSONException if the object is malformed.
-     * @see <a
-     * href="https://developers.google.com/pay/api/android/reference/object#TransactionInfo">TransactionInfo</a>
-     */
-    private static JSONObject getTransactionInfo(String price) throws JSONException {
-        JSONObject transactionInfo = new JSONObject();
-        transactionInfo.put("totalPrice", price);
-        transactionInfo.put("totalPriceStatus", "FINAL");
-        transactionInfo.put("countryCode", ActivityGP.COUNTRY_CODE);
-        transactionInfo.put("currencyCode", ActivityGP.CURRENCY_CODE);
-        transactionInfo.put("checkoutOption", "COMPLETE_IMMEDIATE_PURCHASE");
-
-        return transactionInfo;
-    }
-
-    /**
-     * Information about the merchant requesting payment information
-     *
-     * @return Information about the merchant.
-     * @throws JSONException if the object is malformed.
-     * @see <a
-     * href="https://developers.google.com/pay/api/android/reference/object#MerchantInfo">MerchantInfo</a>
-     */
-    private static JSONObject getMerchantInfo() throws JSONException {
-        return new JSONObject().put("merchantName", "Example Merchant");
-    }
-
-    /**
-     * An object describing information requested in a Google Pay payment sheet
-     *
-     * @return Payment data expected by your app.
-     * @see <a
-     * href="https://developers.google.com/pay/api/android/reference/object#PaymentDataRequest">PaymentDataRequest</a>
-     */
-    public static JSONObject getPaymentDataRequest(long priceCents) {
-
-        final String price = UtilGP.centsToString(priceCents);
-
-        try {
-            JSONObject paymentDataRequest = UtilGP.getBaseRequest();
-            paymentDataRequest.put("allowedPaymentMethods", new JSONArray().put(UtilGP.getCardPaymentMethod()));
-            paymentDataRequest.put("transactionInfo", UtilGP.getTransactionInfo(price));
-            paymentDataRequest.put("merchantInfo", UtilGP.getMerchantInfo());
-
-      /* An optional shipping address requirement is a top-level property of the PaymentDataRequest
-      JSON object. */
-            paymentDataRequest.put("shippingAddressRequired", false);
-//      JSONObject shippingAddressParameters = new JSONObject();
-//      shippingAddressParameters.put("phoneNumberRequired", false);
-//      JSONArray allowedCountryCodes = new JSONArray(Constants.SHIPPING_SUPPORTED_COUNTRIES);
-//      shippingAddressParameters.put("allowedCountryCodes", allowedCountryCodes);
-//      paymentDataRequest.put("shippingAddressParameters", shippingAddressParameters);
-
-            return paymentDataRequest;
-
-        } catch (JSONException e) {
-            return null;
-        }
-    }
-
-    /**
-     * Converts cents to a string format accepted by {@link UtilGP#getPaymentDataRequest}.
-     *
-     * @param cents value of the price in cents.
-     */
-    public static String centsToString(long cents) {
-        return new BigDecimal(cents)
-                .divide(CENTS_IN_A_UNIT, RoundingMode.HALF_EVEN)
-                .setScale(2, RoundingMode.HALF_EVEN)
-                .toString();
-    }
-}
