+++ paymentxdk/.gradle/config.properties
@@ -1,0 +1,2 @@
+#Fri Jun 07 16:46:04 SGT 2024
+java.home=D\:\\Program Files\\Android\\Android Studio 1\\jbr
