+++ paymentxdk/src/main/java/com/molpay/molpayxdk/googlepay/Helper/AlgorithmHelper.java
