+++ paymentxdk/src/main/res/drawable/check_circle.xml
