+++ paymentxdk/src/main/res/layout/activity_googlepay.xml
