+++ paymentxdk/src/main/res/layout/activity_payment.xml
