+++ paymentxdk/src/main/res/layout/activity_web.xml
