+++ paymentxdk/src/main/res/menu/menu_molpay.xml
