+++ paymentxdk/src/main/res/values/dimens.xml
