+++ paymentxdk/src/main/res/values/strings.xml
